+++ layers/CMSSchema/packages/custompost-mutations/src/MutationResolvers/PayloadableUpdateGenericCustomPostMutationResolver.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPCMSSchema\CustomPostMutations\MutationResolvers;
+
+use PoPCMSSchema\CustomPostMutations\MutationResolvers\PayloadableUpdateCustomPostMutationResolverTrait;
+
+class PayloadableUpdateGenericCustomPostMutationResolver extends AbstractCreateUpdateGenericCustomPostMutationResolver
+{
+    use PayloadableUpdateCustomPostMutationResolverTrait;
+}