+++ layers/CMSSchema/packages/custompost-mutations/src/TypeResolvers/ObjectType/GenericCustomPostUpdateMutationPayloadObjectTypeResolver.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPCMSSchema\CustomPostMutations\TypeResolvers\ObjectType;
+
+class GenericCustomPostUpdateMutationPayloadObjectTypeResolver extends AbstractGenericCustomPostMutationPayloadObjectTypeResolver
+{
+    public function getTypeName(): string
+    {
+        return 'GenericCustomPostUpdateMutationPayload';
+    }
+
+    public function getTypeDescription(): ?string
+    {
+        return $this->__('Payload of executing an update nested mutation on a generic custom post', 'custompost-mutations');
+    }
+}