+++ layers/CMSSchema/packages/post-mutations/src/FieldResolvers/ObjectType/PostObjectTypeFieldResolver.php
@@ -12,11 +12,8 @@ use PoPCMSSchema\PostMutations\MutationResolvers\UpdatePostMutationResolver;
 use PoPCMSSchema\PostMutations\TypeResolvers\InputObjectType\PostUpdateInputObjectTypeResolver;
 use PoPCMSSchema\PostMutations\TypeResolvers\ObjectType\PostUpdateMutationPayloadObjectTypeResolver;
 use PoPCMSSchema\Posts\TypeResolvers\ObjectType\PostObjectTypeResolver;
-use PoPCMSSchema\UserState\Checkpoints\UserLoggedInCheckpoint;
 use PoP\ComponentModel\App;
-use PoP\ComponentModel\Checkpoints\CheckpointInterface;
 use PoP\ComponentModel\MutationResolvers\MutationResolverInterface;
-use PoP\ComponentModel\QueryResolution\FieldDataAccessorInterface;
 use PoP\ComponentModel\TypeResolvers\ConcreteTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\InputTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\ObjectType\ObjectTypeResolverInterface;
@@ -28,7 +25,6 @@ class PostObjectTypeFieldResolver extends AbstractCustomPostObjectTypeFieldResol
     private ?UpdatePostMutationResolver $updatePostMutationResolver = null;
     private ?PayloadableUpdatePostMutationResolver $payloadableUpdatePostMutationResolver = null;
     private ?PostUpdateInputObjectTypeResolver $postUpdateInputObjectTypeResolver = null;
-    private ?UserLoggedInCheckpoint $userLoggedInCheckpoint = null;
 
     final public function setPostObjectTypeResolver(PostObjectTypeResolver $postObjectTypeResolver): void
     {
@@ -95,19 +91,6 @@ class PostObjectTypeFieldResolver extends AbstractCustomPostObjectTypeFieldResol
         }
         return $this->postUpdateInputObjectTypeResolver;
     }
-    final public function setUserLoggedInCheckpoint(UserLoggedInCheckpoint $userLoggedInCheckpoint): void
-    {
-        $this->userLoggedInCheckpoint = $userLoggedInCheckpoint;
-    }
-    final protected function getUserLoggedInCheckpoint(): UserLoggedInCheckpoint
-    {
-        if ($this->userLoggedInCheckpoint === null) {
-            /** @var UserLoggedInCheckpoint */
-            $userLoggedInCheckpoint = $this->instanceManager->getInstance(UserLoggedInCheckpoint::class);
-            $this->userLoggedInCheckpoint = $userLoggedInCheckpoint;
-        }
-        return $this->userLoggedInCheckpoint;
-    }
 
     /**
      * @return array<class-string<ObjectTypeResolverInterface>>
@@ -164,40 +147,5 @@ class PostObjectTypeFieldResolver extends AbstractCustomPostObjectTypeFieldResol
                 : $this->getPostObjectTypeResolver(),
             default => parent::getFieldTypeResolver($objectTypeResolver, $fieldName),
         };
-    }
-
-    /**
-     * @return CheckpointInterface[]
-     */
-    public function getValidationCheckpoints(
-        ObjectTypeResolverInterface $objectTypeResolver,
-        FieldDataAccessorInterface $fieldDataAccessor,
-        object $object,
-    ): array {
-        $validationCheckpoints = parent::getValidationCheckpoints(
-            $objectTypeResolver,
-            $fieldDataAccessor,
-            $object,
-        );
-
-        /**
-         * For Payloadable: The "User Logged-in" checkpoint validation is not added,
-         * instead this validation is executed inside the mutation, so the error
-         * shows up in the Payload
-         *
-         * @var CustomPostMutationsModuleConfiguration
-         */
-        $moduleConfiguration = App::getModule(CustomPostMutationsModule::class)->getConfiguration();
-        $usePayloadableCustomPostMutations = $moduleConfiguration->usePayloadableCustomPostMutations();
-        if ($usePayloadableCustomPostMutations) {
-            return $validationCheckpoints;
-        }
-
-        switch ($fieldDataAccessor->getFieldName()) {
-            case 'update':
-                $validationCheckpoints[] = $this->getUserLoggedInCheckpoint();
-                break;
-        }
-        return $validationCheckpoints;
     }
 }